vertechie_be/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# AWS
boto3==1.34.25

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Background Tasks
celery==5.3.6
flower==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1
python-slugify==8.0.1
aiosqlite==0.19.0

